Async AdyenCheckout, better redirecting (#10)
* Create AdyenCheckout with async/await.

Small caveat, componentDid... does not not catch async exceptions. This
means that any exceptions thrown here are only shown in the developer
console and do not cause React to stop.

* Split up session and orderRef in state.

It is probably better for a tuple to be two separate variables.

* Use history.replace for redirecting.

In react-router v4 redirection using `withRouter` and the `history` prop
still are a valid approach. From v5 the `<Navigate>` component or
`useNavigate` should be used.